api_server/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv==1.0.1
PyJWT
python-multipart
